# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Capacity & Spend Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /functions:
    get:
      operationId: listFunctions
      tags: [functions]
      summary: List all business functions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BusinessFunction" }
    post:
      operationId: createFunction
      tags: [functions]
      summary: Create a business function (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BusinessFunctionInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BusinessFunction" }
  /functions/{id}:
    patch:
      operationId: updateFunction
      tags: [functions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BusinessFunctionInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BusinessFunction" }
    delete:
      operationId: deleteFunction
      tags: [functions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /fx-rates:
    get:
      operationId: listFxRates
      tags: [fxRates]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FxRate" }
    post:
      operationId: upsertFxRate
      tags: [fxRates]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FxRateInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FxRate" }
  /fx-rates/{currency}:
    delete:
      operationId: deleteFxRate
      tags: [fxRates]
      parameters:
        - { name: currency, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: portfolio, in: query, schema: { type: string } }
        - { name: phase, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
  /demands:
    get:
      operationId: listDemands
      tags: [demands]
      parameters:
        - { name: projectId, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string } }
        - { name: skill, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Demand" }
    post:
      operationId: createDemand
      tags: [demands]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DemandInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Demand" }
  /demands/{id}:
    patch:
      operationId: updateDemand
      tags: [demands]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DemandInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Demand" }
    delete:
      operationId: deleteDemand
      tags: [demands]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    BusinessFunction:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
    BusinessFunctionInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
    Project:
      type: object
      required:
        - id
        - name
        - code
        - portfolio
        - startDate
        - endDate
        - phase
        - budgetOwner
        - status
        - approvedBudget
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        portfolio: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        phase: { type: string, enum: [Opex, Capex, Run, Change] }
        budgetOwner: { type: string }
        status: { type: string, enum: [Planned, Active, On Hold, Completed, Cancelled] }
        approvedBudget: { type: number }
        functionId: { type: integer, nullable: true }
        functionName: { type: string, nullable: true }
    ProjectInput:
      type: object
      required:
        - name
        - code
        - portfolio
        - startDate
        - endDate
        - phase
        - budgetOwner
        - status
        - approvedBudget
      properties:
        name: { type: string }
        code: { type: string }
        portfolio: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        phase: { type: string, enum: [Opex, Capex, Run, Change] }
        budgetOwner: { type: string }
        status: { type: string, enum: [Planned, Active, On Hold, Completed, Cancelled] }
        approvedBudget: { type: number }
        functionId: { type: integer, nullable: true }
...
    Demand:
      type: object
      required:
        - id
        - projectId
        - projectName
        - skillRequired
        - roleRequired
        - startMonth
        - endMonth
        - requiredFte
        - priority
        - status
      properties:
        id: { type: integer }
        projectId: { type: integer }
        projectName: { type: string }
        skillRequired: { type: string }
        roleRequired: { type: string }
        startMonth: { type: string, description: "YYYY-MM" }
        endMonth: { type: string, description: "YYYY-MM" }
        requiredFte: { type: number }
        priority: { type: string, enum: [Low, Medium, High, Critical] }
        status: { type: string, enum: [Draft, Submitted, Approved, Fulfilled, Rejected] }
    DemandInput:
      type: object
      required:
        - projectId
        - skillRequired
        - roleRequired
        - startMonth
        - endMonth
        - requiredFte
        - priority
        - status
      properties:
        projectId: { type: integer }
        skillRequired: { type: string }
        roleRequired: { type: string }
        startMonth: { type: string }
        endMonth: { type: string }
        requiredFte: { type: number }
        priority: { type: string, enum: [Low, Medium, High, Critical] }
        status: { type: string, enum: [Draft, Submitted, Approved, Fulfilled, Rejected] }
...
    FxRate:
      type: object
      required: [currency, rateToUsd, updatedAt]
      properties:
        currency: { type: string }
        rateToUsd: { type: number }
        updatedAt: { type: string, format: date-time }
    FxRateInput:
      type: object
      required: [currency, rateToUsd]
      properties:
        currency: { type: string }
        rateToUsd: { type: number }
...